+++ ApiMyList/ApiMyList/Repository/IRepositoryProducts.cs
@@ -10,9 +10,9 @@ namespace ApiMyList.Repository
     {
         List<Product> GetProductos(int idLista);
         Product GetProducto(int idProducto);
-        void InsertarProducto(int IdLista, String Nombre, int Cantidad, float Precio, bool chekced);
+        void InsertarProducto(int IdLista, String Nombre, int Cantidad, double Precio, bool chekced);
         void EliminarProducto(int id, int IdLista);
-        void ModificarProducto(int id, int IdLista, String Nombre, int Cantidad, float Precio, bool chekced);
+        void ModificarProducto(int id, int IdLista, String Nombre, int Cantidad, double Precio, bool chekced);
 
         void check(int id, bool check);
